% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{cli_show_user}
\alias{cli_show_user}
\title{User Information Report}
\usage{
cli_show_user(.con)
}
\arguments{
\item{.con}{A valid \code{DBI} connection (DuckDB / MotherDuck).}
}
\value{
This function doesn't return any value. It generates a formatted user report with the current user's
name and role as output.
}
\description{
This function generates a report that shows the current user and their assigned role within the database.
It queries the database to retrieve the current user using \code{current_user()} and the current role using
\code{current_role()}. The output is displayed in a clear and formatted manner, with the user name and role
listed in an unordered list.
}
\keyword{internal}
