% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_fns.R
\name{validate_and_print_database_location}
\alias{validate_and_print_database_location}
\title{Validate and print your database location}
\usage{
validate_and_print_database_location(.con)
}
\arguments{
\item{.con}{A valid \code{DBI} connection (DuckDB / MotherDuck).}
}
\value{
Prints a message with the database file location.
}
\description{
Internal function to help validate your local database location.
}
\keyword{internal}
