% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_fns.R
\name{validate_and_print_database_loction}
\alias{validate_and_print_database_loction}
\title{validate and Pprint your database location}
\usage{
validate_and_print_database_loction(.con)
}
\arguments{
\item{.con}{A valid \code{DBI} connection (DuckDB / MotherDuck).}
}
\value{
print message
}
\description{
Internal function to help validate your local database location
}
