% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_fns.R
\name{validate_con}
\alias{validate_con}
\title{Validate connection is DuckDB}
\usage{
validate_con(.con)
}
\arguments{
\item{.con}{A valid \code{DBI} connection (DuckDB / MotherDuck).}
}
\value{
logical value or error message
}
\description{
Validates that your connection object is a DuckDB connection
}
\examples{
\dontrun{
con <- DBI::dbConnect(duckdb::duckdb())
validate_duckdb_con(con)
}
}
