% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_fns.R
\name{validate_md_connection_status}
\alias{validate_md_connection_status}
\title{Validate Mother Duck Connection Status}
\usage{
validate_md_connection_status(.con, return_type = "msg")
}
\arguments{
\item{.con}{A valid \code{DBI} connection (DuckDB / MotherDuck).}

\item{return_type}{return message or logical value of connection status}
}
\value{
logical value or warning message
}
\description{
Validates if you are successfully connected to motherduck database and will return either a logical value or print a message
}
\examples{
\dontrun{
con <- DBI::dbConnect(duckdb::duckdb())
validate_md_connection_status(con)
}
}
\seealso{
\code{\link[=connect_to_motherduck]{connect_to_motherduck()}}
}
