% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-db.R
\name{validate_database_schema_exists}
\alias{validate_database_schema_exists}
\title{Check if a Schema Exists in a Database}
\usage{
validate_database_schema_exists(.con, database_name, target_schema)
}
\arguments{
\item{.con}{A valid \code{DBI} connection (DuckDB / MotherDuck).}

\item{database_name}{The database name to check.}

\item{target_schema}{The schema name to check within the database.}
}
\value{
Logical \code{TRUE} if the schema exists, \code{FALSE} otherwise.
}
\description{
Checks whether a schema with the specified name exists within a given database.
}
\details{
This function queries the information schema to check if the specified
schema exists within the given database.
}
\examples{
\dontrun{
con <- DBI::dbConnect(duckdb::duckdb())
validate_database_schema_exists(con, "memory", "main")
}

}
